+++ CHANGES.md
@@ -46,6 +46,7 @@ Release Notes.
 * Add `Message Queue Avg Consuming Latency` metric for MQ consuming service and endpoint.
 * Support `-Inf` as bucket in the meter system.
 * Fix setting wrong field when combining `Event`s.
+* Support search browser service.
 
 #### UI